Attempt to split models.py into files
