+++ DataRepo/models/tracerlabeledclass.py
@@ -1,0 +1,27 @@
+# abstract class for models which are "labeled" with tracers; because we have
+# not normalized column names, this is not a Django Abstract base class
+# (https://docs.djangoproject.com/en/3.1/topics/db/models/#abstract-base-classes).
+# This simply shares some configured variables/values.
+class TracerLabeledClass:
+    # choice specifications
+    CARBON = "C"
+    NITROGEN = "N"
+    HYDROGEN = "H"
+    OXYGEN = "O"
+    SULFUR = "S"
+    TRACER_LABELED_ELEMENT_CHOICES = [
+        (CARBON, "Carbon"),
+        (NITROGEN, "Nitrogen"),
+        (HYDROGEN, "Hydrogen"),
+        (OXYGEN, "Oxygen"),
+        (SULFUR, "Sulfur"),
+    ]
+
+    MAX_LABELED_ATOMS = 20
+
+    @classmethod
+    def tracer_labeled_elements_list(cls):
+        tracer_element_list = []
+        for idx in cls.TRACER_LABELED_ELEMENT_CHOICES:
+            tracer_element_list.append(idx[0])
+        return tracer_element_list
